Restructured Project to Allow for Absolute Imports
With old project structure, absolute imports were impossible.
Restructured project with new pipd directory, leaving scripts at the
project root, allowing for absoulte imports for packages and modules.
